+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- **BREAKING**: Renamed `FromRaw`, `IntoRaw` and `DerefRaw` to `FromStorage`, `IntoStorage` and `DerefStorage`
 - Added nested bitfield support, used by specifying a field's type as `nested T`
 
 ## 0.4.0